Refactor heavy TranslationResult and RuntimeAsset header file
