+++ Gems/ScriptCanvas/Code/Editor/Framework/ScriptCanvasGraphUtilities.h
@@ -13,6 +13,7 @@
 #include <Editor/Framework/ScriptCanvasReporter.h>
 #include <ScriptCanvas/Assets/ScriptCanvasFileHandling.h>
 #include <ScriptCanvas/Core/Core.h>
+#include <ScriptCanvas/Translation/TranslationResult.h>
 
 namespace ScriptCanvas
 {
