+++ Gems/ScriptCanvas/Code/Editor/View/Windows/MainWindow.cpp
@@ -147,6 +147,7 @@
 
 #include <Editor/Assets/ScriptCanvasAssetHelpers.h>
 #include <ScriptCanvas/Asset/AssetDescription.h>
+#include <ScriptCanvas/Asset/SubgraphInterfaceAsset.h>
 #include <ScriptCanvas/Components/EditorScriptCanvasComponent.h>
