+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Translation/TranslationResult.h
@@ -11,6 +11,7 @@
 #include <AzCore/Asset/AssetCommon.h>
 #include <AzCore/Script/ScriptAsset.h>
 #include <AzCore/std/any.h>
+#include <ScriptCanvas/Asset/RuntimeInputs.h>
 #include <ScriptCanvas/Core/Core.h>
 #include <ScriptCanvas/Core/SubgraphInterface.h>
 #include <ScriptCanvas/Data/Data.h>
@@ -18,23 +19,8 @@
 #include <ScriptCanvas/Grammar/DebugMap.h>
 #include <ScriptCanvas/Grammar/PrimitivesDeclarations.h>
 
-namespace AZ
-{
-    class EntityId;
-}
-
 namespace ScriptCanvas
 {
-    class Datum;
-    class Nodeable;
-
-    struct VariableId;
-
-    namespace Grammar
-    {
-        class AbstractCodeModel;
-    }
-
     namespace Translation
     {
         enum TargetFlags
@@ -42,40 +28,6 @@ namespace ScriptCanvas
             Lua = 1 << 0,
             Cpp = 1 << 1,
             Hpp = 1 << 2,
-        };
-
-        // information required at runtime begin execution of the compiled graph from the host 
-        struct RuntimeInputs
-        {
-            AZ_TYPE_INFO(RuntimeInputs, "{CFF0820B-EE0D-4E02-B847-2B295DD5B5CF}");
-            AZ_CLASS_ALLOCATOR(RuntimeInputs, AZ::SystemAllocator, 0);
-
-            static void Reflect(AZ::ReflectContext* reflectContext);
-
-            Grammar::ExecutionStateSelection m_executionSelection = Grammar::ExecutionStateSelection::InterpretedPure;
-            
-            AZStd::vector<Nodeable*> m_nodeables;
-            // \todo Datum should be able to be cut by now, replaced by AZStd::any (and maybe TypedNullPointer if necessary)
-            AZStd::vector<AZStd::pair<VariableId, Datum>> m_variables;
-
-            // either the entityId was a (member) variable in the source graph, or it got promoted to one during parsing
-            AZStd::vector<AZStd::pair<VariableId, Data::EntityIDType>> m_entityIds;
-
-            // Statics required for internal, local values that need non-code constructible initialization,
-            // when the system can't pass in the input from C++.
-            AZStd::vector<AZStd::pair<VariableId, AZStd::any>> m_staticVariables;
-
-            RuntimeInputs() = default;
-            RuntimeInputs(const RuntimeInputs&) = default;
-            RuntimeInputs(RuntimeInputs&&);
-            ~RuntimeInputs() = default;
-
-            void CopyFrom(const Grammar::ParsedRuntimeInputs& inputs);
-
-            size_t GetConstructorParameterCount() const;
-
-            RuntimeInputs& operator=(const RuntimeInputs&) = default;
-            RuntimeInputs& operator=(RuntimeInputs&&);
         };
 
         struct TargetResult
