+++ Gems/ScriptCanvas/Code/Tests/ScriptCanvasBuilderTests.cpp
@@ -6,8 +6,8 @@
  *
  */
 
-
 #include <AssetBuilderSDK/AssetBuilderSDK.h>
+#include <AssetBuilderSDK/SerializationDependencies.h>
 #include <AzCore/Asset/AssetManager.h>
 #include <AzCore/Asset/AssetSerializer.h>
 #include <AzCore/Component/Component.h>
@@ -18,8 +18,8 @@
 #include <AzTest/AzTest.h>
 #include <Builder/ScriptCanvasBuilderWorker.h>
 
-#include "ScriptCanvas/Asset/RuntimeAsset.h"
-#include "AssetBuilderSDK/SerializationDependencies.h"
+#include <ScriptCanvas/Core/GraphData.h>
+#include <ScriptCanvas/Asset/RuntimeAsset.h>
 
 struct MockAsset
     : public AZ::Data::AssetData
